requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
line-bot-sdk==3.11.0
requests==2.32.3
python-multipart==0.0.9
